requirements_minimal.txt
# Minimal requirements - NO TensorFlow/PyTorch dependencies
qdrant-client==1.7.0
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
streamlit==1.28.0
pandas==2.1.3

# For development/testing only (optional)
# openai==1.3.0
# anthropic>=0.5.0